Fix two enum-cast warnings

+++ libr/anal/rpnesil.c
@@ -96,7 +96,7 @@ R_API int r_anal_esil_set_op (RAnalEsil *esil, const char *op, RAnalEsilOp code)
 		return R_FALSE;
 	}
 	h = sdb_itoa (sdb_hash (op), t, 16);
-	sdb_num_set (esil->ops, h, (ut64)code, 0);
+	sdb_num_set (esil->ops, h, (ut64)(size_t)code, 0);
 	if (!sdb_num_exists (esil->ops, h))
 		eprintf ("can't set esil-op %s\n", op);
 	return R_TRUE;
@@ -2241,7 +2241,7 @@ static int iscommand (RAnalEsil *esil, const char *word, RAnalEsilOp *op) {
 	char *h;
 	h = sdb_itoa (sdb_hash (word), t, 16);
 	if (sdb_num_exists (esil->ops, h)) {
-		*op = (RAnalEsilOp) sdb_num_get (esil->ops, h, 0);
+		*op = (RAnalEsilOp) (size_t)sdb_num_get (esil->ops, h, 0);
 		return R_TRUE;
 	}
 	return R_FALSE;
